requirements.txt
djangorestframework>=3.14.0
Django>=5.0.7
django-debug-toolbar>=4.2.0
django-filter>=24.3
django-environ>=0.11.2
psycopg2-binary>=2.9.5

celery>=5.3.6
redis>=5.0.1
flower>=2.0.1

pytest>=7.1.3
pytest-django>=4.8.0
